+++ src/PHPUnit/Command/index.ts
@@ -1,1 +1,1 @@
-export * from './Command';
+export * from './CommandBuilder';